fixed styling issue with main header now using webpack-dev-server for local server
