+++ package.json
@@ -31,7 +31,6 @@
     "polished": "1.9.2",
     "popper.js": "1.12.9",
     "prop-types": "15.6.0",
-    "raw-loader": "0.5.1",
     "react": "16.2.0",
     "react-dom": "16.2.0",
     "react-dropzone": "4.2.8",
@@ -80,7 +79,7 @@
     "eslint-config-rapid7": "2.11.1",
     "eslint-plugin-react": "7.6.1",
     "extract-text-webpack-plugin": "3.0.2",
-    "file-loader": "1.1.6",
+    "file-loader": "1.1.11",
     "flow-bin": "0.66.0",
     "glob-array": "1.0.0",
     "html-webpack-plugin": "2.30.1",
@@ -95,6 +94,7 @@
     "precss": "3.1.0",
     "prettier-eslint-cli": "4.7.1",
     "raf": "3.4.0",
+    "raw-loader": "0.5.1",
     "react-test-renderer": "16.2.0",
     "redux-freeze": "0.1.5",
     "resolve-url-loader": "2.2.1",
@@ -103,8 +103,7 @@
     "string.format": "1.3.1",
     "style-loader": "0.20.1",
     "webpack": "3.10.0",
-    "webpack-clean-obsolete-chunks": "0.3.0",
-    "webpack-livereload-plugin": "1.0.0"
+    "webpack-dev-server": "2.11.1"
   },
   "engines": {
     "node": ">=4.4.3"
@@ -131,22 +130,18 @@
     "url": "git://github.com/ryanzec/underpin-ui.git"
   },
   "scripts": {
-    "build:developer": "echo \"Building code...\" && npm run clean && npm run copy-static-assets && npm run webpack",
-    "build:production": "echo \"Building code...\" && npm run clean && npm run copy-static-assets && npm run webpack:production",
-    "build:watch": "echo \"Building code...\" && npm run clean && npm run copy-static-assets && npm run webpack:watch",
-    "clean": "echo \"Cleaning out build directory...\" && rm -rf ./web/build",
-    "copy-static-assets": "echo \"Copying static assets...\" && mkdir -p ./web/build/images && cp -R ./web/app/images/ ./web/build/images",
-    "lint": "echo \"Linting javascript code...\" && ./node_modules/.bin/eslint --config prod.eslintrc src/ test/",
-    "lint:fix": "echo \"Auto formating javascript code...\" && ./node_modules/.bin/prettier-eslint --write \"src/**/*.js\" \"test/**/*.js\"",
+    "build:production": "yarn clean && yarn && NODE_ENV=production WEBPACK_PRODUCTION_BUILD=1 ./node_modules/.bin/webpack --progress --config ./webpack/webpack.prod.config.js",
+    "dev": "NODE_ENV=development node ./node_modules/.bin/webpack-dev-server --colors --progress --profile --config ./webpack/webpack.dev.config.js",
+    "clean": "rm -rf ./web/build",
+    "copy-static-assets": "mkdir -p ./web/build/images && cp -R ./web/app/images/ ./web/build/images",
+    "lint": "./node_modules/.bin/eslint --config prod.eslintrc src/ test/",
+    "lint:fix": "./node_modules/.bin/prettier-eslint --write \"src/**/*.js\" \"test/**/*.js\"",
     "precommit": "lint-staged && yarn lint",
     "prepush": "yarn lint && yarn test",
     "prepublish": "rm -rf ./lib && mkdir -p ./lib/styles && ./node_modules/babel-cli/bin/babel.js src --out-dir lib && cp -r ./src/styles/ ./lib/styles",
     "test": "NODE_PATH=. NODE_ENV=test ./node_modules/.bin/ava --verbose",
     "test:watch": "NODE_PATH=. NODE_ENV=test ./node_modules/.bin/ava --verbose --watch",
-    "test:update": "NODE_PATH=. NODE_ENV=test ./node_modules/.bin/ava --verbose --update-snapshots",
-    "webpack": "yarn clean && yarn && NODE_ENV=development ./node_modules/.bin/webpack --progress --config webpack.config.js",
-    "webpack:production": "yarn clean && yarn && NODE_ENV=production WEBPACK_PRODUCTION_BUILD=1 ./node_modules/.bin/webpack --progress --config webpack.config.js",
-    "webpack:watch": "yarn clean && yarn && NODE_ENV=development ./node_modules/.bin/webpack --progress --config webpack.config.js --watch"
+    "test:update": "NODE_PATH=. NODE_ENV=test ./node_modules/.bin/ava --verbose --update-snapshots"
   },
   "version": "0.1.0"
 }
